+++ src/main/java/com/jasik/momsnaggingapi/domain/diary/controller/DiaryController.java
@@ -1,0 +1,7 @@
+package com.jasik.momsnaggingapi.domain.diary.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class DiaryController {
+}
